+++ slider/src/network/address.cpp
@@ -1,0 +1,30 @@
+#include "address.h"
+
+using namespace Net;
+
+MacAddress::MacAddress() {}
+
+MacAddress::MacAddress(std::array<uint8_t, 6> address)
+    : m_Address(address) {}
+
+MacAddress::MacAddress(const uint8_t *address)
+{
+    std::copy(address, address + 6, m_Address.begin());
+}
+
+void MacAddress::CopyTo(uint8_t *dest) const
+{
+    std::copy(m_Address.begin(), m_Address.end(), dest);
+}
+
+size_t MacAddress::printTo(Print &p) const
+{
+    size_t n = 0;
+    for (int i = 0; i < 6; i++)
+    {
+        n += p.printf("%02x", m_Address[i]);
+        if (i < 6 - 1)
+            n += p.print(':');
+    }
+    return n;
+}
