+++ slider/src/test/network/netApp.cpp
@@ -2,7 +2,7 @@
 #include "src/app/appConfig.h"
 #include "src/debug.h"
 #include "messageHandler.h"
-#include "address.h"
+#include "src/network/address.h"
 #include "src/network/wifi.h"
 
 using namespace Slider;
